% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_to_y.R
\name{add_noise_to_y}
\alias{add_noise_to_y}
\title{Add Gaussian Noise to y with Optional Heteroscedasticity}
\usage{
add_noise_to_y(df, noise_level, heteroscedastic = FALSE)
}
\arguments{
\item{df}{A data frame with at least two columns: 'x' and 'y'.}

\item{noise_level}{Character, specifies the intensity of the noise: 'low' or 'high'.}

\item{heteroscedastic}{Logical, if TRUE, adds heteroscedastic noise, otherwise homoscedastic noise.}
}
\value{
A data frame with the same 'x' values and adjusted 'y' values with added Gaussian noise.
}
\description{
This function adds Gaussian noise to the 'y' values of a data frame based on a specified
noise level ('low' or 'high'). Optionally, it can add heteroscedastic noise where the
variance of the noise increases with the value of 'x'.
}
\examples{
df <- data.frame(x = 1:100, y = sin(1:100) * 100)
noisy_df_low <- add_noise_to_y(df, noise_level = "low", heteroscedastic = FALSE)
noisy_df_high_hetero <- add_noise_to_y(df, noise_level = "high", heteroscedastic = TRUE)
plot(noisy_df_low$x, noisy_df_low$y, type = "l", col = "blue", main = "Noisy Sine Wave")
lines(noisy_df_high_hetero$x, noisy_df_high_hetero$y, col = "red")

}
