% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-const.R
\docType{data}
\name{anova_text}
\alias{anova_text}
\title{Analysis of Variance (ANOVA) Exercise Scenarios}
\format{
An object of class \code{list} of length 20.
}
\usage{
anova_text
}
\description{
This list contains 20 scenarios designed for practicing Analysis of Variance (ANOVA).
Each scenario includes categories (as factor levels) and a specific prompt that outlines the
experiment or study context where ANOVA could be applied to determine if there are statistically
significant differences across group means.
}
\keyword{datasets}
