% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_fit_nonuniform_text.R
\docType{data}
\name{chisq_fit_nonuniform_text}
\alias{chisq_fit_nonuniform_text}
\title{Chi-square Goodness of Fit Test Scenarios for Non-uniform Distributions}
\format{
An object of class \code{list} of length 50.
}
\usage{
chisq_fit_nonuniform_text
}
\description{
Contains lists of category pairs with descriptions and prompts for conducting
Chi-square tests of goodness of fit, where different groups are expected to have different
distributions. This includes comparisons of subgroups within broader categories.
}
\keyword{datasets}
