% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-const.R
\docType{data}
\name{chisq_fit_uniform_text}
\alias{chisq_fit_uniform_text}
\title{Chi-square Goodness of Fit Test Scenarios for Uniform Distributions}
\format{
An object of class \code{list} of length 67.
}
\usage{
chisq_fit_uniform_text
}
\description{
Contains lists of categories and their corresponding prompts for conducting
Chi-square tests of goodness of fit, assuming uniform distributions across categories.
Each element is designed for different practical scenarios where the hypothesis is
that each category is equally likely.
}
\keyword{datasets}
