% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-const.R
\docType{data}
\name{chisq_independence_text}
\alias{chisq_independence_text}
\title{Chi-square Tests of Independence Scenarios}
\format{
An object of class \code{list} of length 20.
}
\usage{
chisq_independence_text
}
\description{
Contains lists of categories for Chi-square tests of independence, detailing
various combinations of categorical data to explore relationships between different groups.
These tests are designed to determine if there is a statistically significant association
between two categorical variables.
}
\keyword{datasets}
