% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_youtube.R
\name{embed_youtube}
\alias{embed_youtube}
\title{Embed a YouTube Video}
\usage{
embed_youtube(
  youtube.id,
  start = "0",
  title = "YouTube video player",
  width = "100\%",
  height = "600"
)
}
\arguments{
\item{youtube.id}{A string specifying the YouTube video ID.}

\item{start}{Start time of the video in seconds (default is 0).}

\item{title}{Title of the iframe (default is "YouTube video player").}

\item{width}{Width of the iframe (default is "100\%").}

\item{height}{Height of the iframe (default is 600 pixels).}
}
\value{
HTML iframe code for the video player. This is printed to the console but not returned.
}
\description{
This function creates an HTML iframe to embed a YouTube video. It allows customization of the start time, title, width, and height of the video player.
}
\examples{
embed_youtube("dQw4w9WgXcQ", start = "10", title = "Sample Video", width = "560", height = "315")
}
