% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_data_as_link.R
\name{encode_data_as_link}
\alias{encode_data_as_link}
\title{Encode Data Frame as Downloadable CSV Link}
\usage{
encode_data_as_link(data, file_name = "download.csv")
}
\arguments{
\item{data}{A data frame to be encoded and downloaded.}

\item{file_name}{The desired name of the downloadable CSV file.}
}
\value{
An HTML string containing a download link which, when clicked,
triggers the download of the data frame as a CSV file with the given file name.
}
\description{
This function takes a data frame and encodes it as a Base64 CSV string,
then returns an HTML link for downloading the encoded data. The link can be
used in web applications or R Markdown documents to directly download data as CSV.
}
\examples{
df <- data.frame(Name = c("Alice", "Bob"), Age = c(25, 30))
html_link <- encode_data_as_link(df)
cat(html_link)
}
