% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nth_values.R
\name{get_nth_values}
\alias{get_nth_values}
\title{Retrieve nth Largest and nth Smallest Values from a Vector}
\usage{
get_nth_values(vec, n)
}
\arguments{
\item{vec}{A numeric vector from which to find the nth largest and nth smallest values.}

\item{n}{An integer indicating the position from the largest and smallest ends to retrieve.
Must be positive and not greater than the length of the vector.}
}
\value{
A list containing the nth largest and nth smallest values.
}
\description{
This function returns the nth largest and nth smallest values from a given numeric vector.
}
\examples{
vec <- c(10, 20, 30, 40, 50)
get_nth_values(vec, 2) # Returns 2nd largest and 2nd smallest values

}
