% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_all}
\alias{make_all}
\title{Calculate hardness and discrimination for all questions}
\usage{
make_all(d_wide)
}
\arguments{
\item{d_wide}{wide-format marks df with each question in one column}
}
\value{
data frame of questions, hardness values and item-total correlations
}
\description{
Calculate hardness and discrimination for all questions
}
\examples{
make_all(test_grades)

}
