% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_discrim}
\alias{make_discrim}
\title{Calculate question discrimination (item-total correlation)}
\usage{
make_discrim(d_long)
}
\arguments{
\item{d_long}{long-format marks df as made by `make_long`}
}
\value{
data frame of questions and item-total correlations over students
}
\description{
Calculate question discrimination (item-total correlation)
}
\examples{
test_long <- make_long(test_grades)
make_discrim(test_long)

}
