% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_hardness}
\alias{make_hardness}
\title{Calculate question hardness}
\usage{
make_hardness(d_long)
}
\arguments{
\item{d_long}{long-format marks df as made by `make_long`}
}
\value{
data frame of questions and mean percent of maximum grade over students
}
\description{
Calculate question hardness
}
\examples{
test_long <- make_long(test_grades)
make_hardness(test_long)

}
