% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longer.R
\name{make_long}
\alias{make_long}
\title{Make longer marks df}
\usage{
make_long(marks_wide)
}
\arguments{
\item{marks_wide}{A wider df with one column per question and a column `Total` that is sum of marks for each student}
}
\value{
A longer df with all marks in one column and for each student and question: that student's total, the question number, the mark on that question
}
\description{
Make longer marks df
}
\examples{
make_long(test_grades)
}
