% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{make_stats}
\alias{make_stats}
\title{Compute exam total score statistics}
\usage{
make_stats(d_wide, ...)
}
\arguments{
\item{d_wide}{wide-format marks df with each question in one column}

\item{...}{additional arguments to `quantile`}
}
\value{
named vector of quantiles of exam totals
}
\description{
Compute exam total score statistics
}
\examples{
make_stats(test_grades)
make_stats(test_grades, c(0.4, 0.6))

}
