% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_anova_.R
\name{question_anova}
\alias{question_anova}
\title{Generate ANOVA Question Data and Perform Analysis}
\usage{
question_anova(
  levels = NULL,
  prompt = NULL,
  means = NULL,
  sds = NULL,
  noise_level = sample(c("high", "low"), 1),
  heteroscedasticity = sample(c(TRUE, FALSE), 1),
  folder_path = "examstats",
  filename = NULL,
  n = 250,
  target_p_value = 0.05,
  email = NULL,
  max_iterations = 1e+05
)
}
\arguments{
\item{levels}{A character vector specifying the levels of the factor variable. If NULL, levels are selected randomly from a predefined list.}

\item{prompt}{A character string specifying the prompt for the question. If NULL, a prompt is selected randomly from a predefined list.}

\item{means}{A numeric vector specifying the means for each level. Default is NULL.}

\item{sds}{A numeric vector specifying the standard deviations for each level. Default is NULL.}

\item{noise_level}{A character string specifying the noise level. Can be "high" or "low". Default is a random selection between "high" and "low".}

\item{heteroscedasticity}{A logical value indicating whether to introduce heteroscedasticity. Default is a random selection between TRUE and FALSE.}

\item{folder_path}{A character string specifying the path to the folder where the data will be saved. Default is "examstats".}

\item{filename}{A character string specifying the filename for the data file. If NULL, a filename is selected randomly from a predefined list.}

\item{n}{An integer specifying the number of observations to generate. Default is 250.}

\item{target_p_value}{A numeric value specifying the target p-value for the ANOVA. Default is 0.05.}

\item{email}{A character string specifying the email address for Google Drive upload. Default is NULL.}

\item{max_iterations}{An integer specifying the maximum number of iterations for data generation. Default is 100000.}
}
\value{
A list containing the following elements:
\describe{
\item{levels}{The levels of the factor variable.}
\item{prompt}{The prompt for the question.}
\item{data}{The generated data frame.}
\item{link}{A link to download the data.}
\item{google_link}{A link to the data file on Google Drive, if uploaded.}
\item{anova_result}{The result of the ANOVA.}
\item{anova_summary}{The summary of the ANOVA.}
\item{levene_test}{The result of Levene's Test for homogeneity of variance.}
\item{shapiro_test}{The result of the Shapiro-Wilk Test for normality of residuals.}
}
}
\description{
This function generates data for an ANOVA question based on specified levels,
means, standard deviations, noise level, and heteroscedasticity. It performs
ANOVA, checks for homogeneity of variance using Levene's Test, and checks for
normality of residuals using the Shapiro-Wilk Test. Optionally, it uploads
the data to Google Drive and returns a link.
}
\examples{
question_anova()


}
