% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_chisq_independence.R
\name{question_chisq_independence}
\alias{question_chisq_independence}
\title{Generate a Chi-Square Test of Independence Question}
\usage{
question_chisq_independence(
  levels1 = NULL,
  levels2 = NULL,
  sample_size = NULL,
  target_p_value = 0.05,
  max_iterations = 1000,
  folder_path = "examstats",
  filename = NULL,
  n = 50,
  email = NULL
)
}
\arguments{
\item{levels1}{A vector of categories for the first variable.
If NULL, randomly selected from predefined scenarios.}

\item{levels2}{A vector of categories for the second variable.
If NULL, randomly selected from predefined scenarios.}

\item{sample_size}{The total number of observations to generate for the data simulation.
If NULL, a default size must be assumed within the simulation function.}

\item{target_p_value}{The target p-value for ensuring the data fits the Chi-square test criteria.
This parameter controls the simulation to achieve a dataset that can
realistically represent independence or dependence as needed for educational purposes.}

\item{max_iterations}{The maximum number of iterations to attempt to generate data that meets the target p-value.}

\item{folder_path}{The folder path where the data file will be saved on Google Drive.}

\item{filename}{The name of the file to save. If NULL, the data will not be saved.}

\item{n}{The number of questions to generate. Only relevant if levels1 and levels2 are NULL.}

\item{email}{The email address to which the Google Drive link of the saved data file will be sent.
If NULL, no email will be sent.}
}
\value{
A list containing the category levels for both variables, the prompt for the question,
the simulated data, a contingency table of the data, and the Chi-square test results.
}
\description{
This function dynamically generates a question for a Chi-square test of independence
using either specified levels or randomly selected from predefined scenarios.
It simulates data assuming independence between the two categorical variables,
performs the Chi-square test, and returns the results along with the generated data.
}
\details{
The function can optionally save the simulated data to a specified folder and
email the saved file link using Google Drive integration if provided with an email address.
}
\examples{
question <- question_chisq_independence()
print(question)

}
