% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_chisq_nonuniform.R
\name{question_chisq_nonuniform}
\alias{question_chisq_nonuniform}
\title{Generate a Chi-Square Goodness of Fit Test Question for Non-uniform Distributions}
\usage{
question_chisq_nonuniform(
  description = NULL,
  levels = NULL,
  prompt = NULL,
  probs = NULL,
  n = 50,
  target_p_value = 0.05,
  max_iterations = 1e+05
)
}
\arguments{
\item{description}{Descriptive text for the question scenario.
If NULL, it is automatically filled with a random scenario.}

\item{levels}{A vector of levels (categories) for the test.
If NULL, a random scenario is selected with corresponding levels.}

\item{prompt}{A string representing the prompt for the test question.
If NULL, it is filled with a prompt corresponding to the selected scenario.}

\item{probs}{A vector of probabilities for each level, representing non-uniform distributions.
If NULL, random non-uniform probabilities are generated.}

\item{n}{The number of samples to generate for the category data.}

\item{target_p_value}{The target p-value for ensuring the data fits the intended Chi-square test criteria.}

\item{max_iterations}{The maximum number of iterations to attempt to meet the target p-value.}
}
\value{
A list containing the question description, assigned probabilities, levels, prompt,
simulated data, a frequency table of the data, and the Chi-square test results.
}
\description{
This function dynamically generates a Chi-square goodness of fit test question
for non-uniform distributions based on either supplied parameters or randomly selected
scenarios from predefined lists. It simulates category data based on specified or
non-uniform probabilities, performs the Chi-square test, and returns the results
along with the generated data.
}
\examples{
question <- question_chisq_nonuniform()
print(question)
}
