% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_chisq_uniform.R
\name{question_chisq_uniform}
\alias{question_chisq_uniform}
\title{Generate a Chi-Square Goodness of Fit Test Question}
\usage{
question_chisq_uniform(
  levels = NULL,
  prompt = NULL,
  probs = NULL,
  folder_path = "examstats",
  filename = NULL,
  n = 250,
  target_p_value = 0.05,
  email = NULL,
  max_iterations = 1e+05
)
}
\arguments{
\item{levels}{A vector of levels (categories) for the test.
If NULL, a random scenario from \code{chisq_fit_uniform_text} is selected.}

\item{prompt}{A string representing the prompt for the test question.
If NULL, associated with the randomly selected scenario.}

\item{probs}{A vector of probabilities for each level.
If NULL, assumes uniform distribution across levels.}

\item{folder_path}{Google Drive folder for data.}

\item{filename}{A file name for an encoded csv file.}

\item{n}{The number of samples to generate for the category data.}

\item{target_p_value}{The target p-value for ensuring the data fits the Chi-square test criteria.}

\item{email}{Email associated with the Google Drive account used for sharing data description}

\item{max_iterations}{The maximum number of iterations to attempt to meet the target p-value.}
}
\value{
A list containing the levels, prompt, data, encoded link, and Google Drive links, a frequency table of the data, and the Chi-square test results.
}
\description{
This function dynamically generates a Chi-square goodness of fit test question
based on a given set of levels or randomly selected from predefined scenarios.
It simulates category data based on specified or uniform probabilities, performs
the Chi-square test, and returns the results along with the generated data.
}
\examples{
question <- question_chisq_uniform()
print(question)

}
