% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdrr_io.R
\name{rdrr_io}
\alias{rdrr_io}
\title{Embed R Code Snippets using rdrr.io}
\usage{
rdrr_io(code = "# Your code here.\\n", width = "100\%", height = "600")
}
\arguments{
\item{code}{A string of R code to be embedded. Default is "# Your code here.".}

\item{width}{Width of the iframe (default is "100\%").}

\item{height}{Height of the iframe (default is 600 pixels).}
}
\value{
HTML iframe code for embedding the R code snippet. It prints the iframe
to the console but does not return it conventionally.
}
\description{
This function creates an HTML iframe to embed R code snippets from rdrr.io.
It allows users to share and execute R code directly from a web page. The function
handles URL encoding of the R code to ensure it is correctly interpreted by the web browser.
}
\examples{
rdrr_io(code = "plot(cars)", width = "560", height = "315")
}
