% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catagory_series.R
\name{simulate_category_series}
\alias{simulate_category_series}
\title{Simulate Categorical Series}
\usage{
simulate_category_series(
  levels,
  probs = NULL,
  n = 250,
  target_p_value = 0.05,
  max_iterations = 1000
)
}
\arguments{
\item{levels}{A vector specifying the levels for the categorical variable.}

\item{probs}{A vector specifying the probabilities for each level. If NULL, probabilities are assumed to be uniform.}

\item{n}{The total number of observations to generate.}

\item{target_p_value}{The target p-value for the chi-squared test.}

\item{max_iterations}{The maximum number of iterations to achieve the target p-value.}
}
\value{
A factor representing the generated categorical variable.
}
\description{
This function generates a categorical variable with specified probabilities, adjusting frequencies until the chi-squared test of independence reaches a target p-value.
}
\examples{
levels <- c("A", "B", "C", "D")
target_p_value <- 0.15
result_df <- simulate_category_series(levels,
                                      target_p_value = target_p_value,
                                      probs = c(0.2, 0.4, 0.1, 0.3))
table(result_df)
chisq.test(table(result_df), p = c(0.2, 0.4, 0.1, 0.3))

}
