% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_correlated_series.R
\name{simulate_correlated_series}
\alias{simulate_correlated_series}
\title{Simulate Two Correlated Series}
\usage{
simulate_correlated_series(n, rho)
}
\arguments{
\item{n}{Integer, the number of observations in each series.}

\item{rho}{Numeric, the desired correlation coefficient between the two series.
This value must be between -1 and 1, inclusive.}
}
\value{
A \code{data.frame} with two columns (\code{x} and \code{y}). Each column contains
one of the simulated series. The correlation between these two series approximates
the specified \code{rho}.
}
\description{
This function generates two series of random data that are correlated to a specified degree,
using the Cholesky decomposition method. The function is useful for simulations in studies
where correlated data sets are required to test statistical methods or models.
}
\examples{
# Simulate two series with a correlation coefficient of 0.75
set.seed(123)
data <- simulate_correlated_series(100, 0.75)
print(head(data))

# Check the actual correlation coefficient
cor(data$x, data$y)

}
