% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_independent_catagory_series.R
\name{simulate_independent_catagory_series}
\alias{simulate_independent_catagory_series}
\title{Simulate Independent Categorical Series}
\usage{
simulate_independent_catagory_series(
  levels1,
  levels2,
  sample_size = NULL,
  target_p_value = 0.05,
  max_iterations = 1000
)
}
\arguments{
\item{levels1}{A vector specifying the levels for the first factor.}

\item{levels2}{A vector specifying the levels for the second factor.}

\item{sample_size}{The desired total sample size. If NULL, it will be calculated based on the length of levels1 and levels2.}

\item{target_p_value}{The target p-value for the chi-squared test of independence.}

\item{max_iterations}{The maximum number of iterations to achieve the target p-value.}
}
\value{
A data frame containing the generated factors with the specified independence.
}
\description{
This function generates two categorical factors with independence between them.
}
\examples{
levels1 <- c("Red", "Blue", "Green", "Red", "Purple")
levels2 <- c("Circle", "Square", "Triangle")
sample_size <- 150
target_p_value <- 0.25
result_df <- simulate_independent_catagory_series(levels1 = levels1,
                                                  levels2 = levels2,
                                                  sample_size = sample_size,
                                                  target_p_value = target_p_value)
(tbl <- table(result_df$Factor1, result_df$Factor2))
(test <- chisq.test(tbl))

}
