% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_related_series.R
\name{simulate_related_series}
\alias{simulate_related_series}
\title{Simulate Related Series}
\usage{
simulate_related_series(sample_size, relationship)
}
\arguments{
\item{sample_size}{The desired size of the data sample.}

\item{relationship}{The type of relationship between x and y variables: 'linear', 'linear-log', 'log-log', 'log-linear', or 'quadratic'.}
}
\value{
A data frame containing the generated x and y variables.
}
\description{
This function generates a related series of data points based on specified relationships such as linear, linear-log, log-log, log-linear, or quadratic.
}
\examples{
# Generate a linear relationship
linear_data <- simulate_related_series(sample_size = 100, relationship = "linear")
# Generate a log-linear relationship
log_linear_data <- simulate_related_series(sample_size = 100, relationship = "log-linear")
# Generate a quadratic relationship
quadratic_data <- simulate_related_series(sample_size = 100, relationship = "quadratic")

}
