% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_grades}
\alias{test_grades}
\title{Exam marks for testing}
\format{A data frame with 5 rows and 4 columns:
\describe{
   \item{Q1-Q3}{marks on questions 1 to 3}
   \item{Total}{Total marks for exam}
}}
\usage{
test_grades
}
\description{
A small dataset of made-up exam marks for testing with
}
\keyword{datasets}
