% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_google.R
\name{upload_data_google}
\alias{upload_data_google}
\title{Upload a DataFrame to Google Drive and Get a Shareable Link}
\usage{
upload_data_google(data, folder_path, file_name, email = NULL)
}
\arguments{
\item{data}{A DataFrame to be uploaded.}

\item{folder_path}{A character string specifying the path of the folder on Google Drive.}

\item{file_name}{A character string specifying the name of the file to be created on Google Drive.}

\item{email}{An optional character string specifying the email to use for authentication.
If NULL, default authentication will be used.}
}
\value{
A character string containing the shareable link to the uploaded file.
}
\description{
This function uploads a given DataFrame to a specified folder on Google Drive
and returns a shareable link. If the folder does not exist, it will be created.
Optionally, you can specify an email for authentication.
}
