% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_chisq_nonuniform.R
\name{question_chisq_nonuniform}
\alias{question_chisq_nonuniform}
\title{Generate a Chi-Square Goodness of Fit Test Question for Non-uniform Distributions}
\usage{
question_chisq_nonuniform(
  description = NULL,
  levels = NULL,
  prompt = NULL,
  probs = NULL,
  folder_path = "examstats",
  filename = NULL,
  n = 50,
  target_p_value = 0.05,
  email = NULL,
  max_iterations = 1e+05
)
}
\arguments{
\item{description}{Descriptive text for the question scenario. If NULL, automatically
selects a random predefined scenario.}

\item{levels}{Categories to be tested in the Chi-square test. If NULL, selects
categories based on a random predefined scenario.}

\item{prompt}{Instructional prompt for the test question. If NULL, a suitable prompt
is automatically generated based on the scenario.}

\item{probs}{Probabilities associated with each category, reflecting a non-uniform
distribution. If NULL, generates random non-uniform probabilities.}

\item{folder_path}{Path to the folder where generated data files will be stored.}

\item{filename}{Name of the file to save the generated data. If NULL, the data
is not saved to a file.}

\item{n}{Number of samples to simulate for the categorical data.}

\item{target_p_value}{Desired p-value threshold for the Chi-square test to ensure the
data meets specific statistical criteria.}

\item{email}{Email address to send the link of the generated data file. If NULL,
the link is not sent.}

\item{max_iterations}{Maximum number of iterations to attempt generating data that
meets the target_p_value.}
}
\value{
A list containing the question scenario, probabilities, categories, prompt,
simulated data, encoded link, and Google Drive links, frequency table of the data,
and Chi-square test results.
}
\description{
This function dynamically generates a question for a Chi-square goodness of fit test
tailored to non-uniform distributions. It allows customization through parameters or
automatically generates data based on pre-defined scenarios. The function simulates
categorical data, performs the Chi-square test, and provides detailed output including
test results and the data used.
}
\examples{
question <- question_chisq_nonuniform()
print(question)
}
